requirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altgraph @ file:///AppleInternal/Library/BuildRoots/bcce998f-ff34-11ef-9d34-f2a857e00a32/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
anyio==4.2.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==5.0.1
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.2
black==23.9.1
bleach==6.1.0
cachetools==5.3.2
captum @ git+https://github.com/pytorch/captum.git@39637231c8920e880aea45e1d1d355ca3ac8939c
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
cmake==3.28.3
colpali_engine @ git+https://github.com/illuin-tech/colpali.git@9b8362af15fdfbc98c003e3295110121e2adb235
comm==0.2.1
contourpy==1.2.0
cryptography==41.0.4
cycler==0.12.1
datasets==3.5.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
docopt==0.6.2
errorhandler==2.0.1
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
ffprobe==0.5
filelock==3.13.3
flake8==6.1.0
fonttools==4.49.0
fqdn==1.5.1
frozenlist==1.6.0
fsspec==2024.3.1
future @ file:///AppleInternal/Library/BuildRoots/bcce998f-ff34-11ef-9d34-f2a857e00a32/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
google-auth==2.27.0
google-auth-oauthlib==1.2.0
gspread==6.0.1
httplib2==0.22.0
huggingface-hub==0.30.2
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipykernel==6.29.0
ipython==8.18.1
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
Js2Py==0.74
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.1
jupyterlab==4.0.10
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kagglehub==0.3.12
kiwisolver==1.4.5
macholib @ file:///AppleInternal/Library/BuildRoots/bcce998f-ff34-11ef-9d34-f2a857e00a32/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
MarkupSafe==2.1.3
matplotlib==3.8.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==3.0.2
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.5.9
networkx==3.2.1
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
overrides==7.4.0
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.2
peft==0.15.2
pexpect==4.9.0
pillow==10.2.0
pipwin==0.5.2
platformdirs==3.10.0
pocketsphinx==5.0.3
prometheus-client==0.19.0
prompt-toolkit==3.0.43
propcache==0.3.1
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyarrow==19.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.0
pycparser==2.21
pydub==0.25.1
pyflakes==3.1.0
PyGithub==1.59.1
Pygments==2.17.2
pyjsparser==2.7.1
PyJWT==2.8.0
PyNaCl==1.5.0
pyparsing==3.1.1
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.32.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.17.1
rsa==4.9
safetensors==0.5.3
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
sounddevice==0.4.6
soupsieve==2.5
SpeechRecognition==3.10.1
stack-data==0.6.3
StrEnum==0.4.15
sympy==1.13.1
terminado==0.18.0
threadpoolctl==3.3.0
tinycss2==1.2.1
tokenizers==0.21.1
tomli==2.0.1
torch==2.6.0
torchaudio==2.2.2
torchvision==0.17.2
tornado==6.4
tqdm==4.67.1
traitlets==5.14.1
transformers==4.51.3
types-python-dateutil==2.8.19.20240106
typing_extensions==4.13.2
tzdata==2023.4
tzlocal==5.2
uri-template==1.3.0
urllib3==2.2.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
widgetsnbextension==4.0.9
wrapt==1.15.0
xxhash==3.5.0
yarl==1.20.0
zipp==3.17.0
